producer/requirements.txt
pandas==2.2.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
kafka-python==2.0.2
